+++ resources/views/components/input.blade.php
@@ -1,4 +1,4 @@
-<div class="{{ $fieldWrapperClass }}">
+<div class="{{ $fieldWrapperClass }} @error('formProperties.' .  $key){{ $fieldErrorWrapperClass }}@enderror">
     @include('lara-forms-builder::includes.field-label')
     @if (isset($mode) && ($mode == 'view' || $mode == 'confirm'))
         <div class="lfb-input-wrapper lfb-input-readonly" x-data="{ showSecretValue: false }">
@@ -23,7 +23,11 @@
             ])
             x-data="{ showSecretValue: false }">
             <input wire:key="form-input-component-{{ md5($key) }}"
-                   wire:model.live="formProperties.{{ $key }}"
+                @if(isset($fieldModifier))
+                    wire:model.{{$fieldModifier}}="formProperties.{{ $key }}"
+                @else
+                    wire:model.live="formProperties.{{ $key }}"
+                @endif
                    id="formProperties-{{ $key }}"
                    name="formProperties.{{ $key }}"
                    type="{{ $inputType ?? 'text' }}"