+++ resources/views/components/select.blade.php
@@ -1,4 +1,4 @@
-<div class="{{ $fieldWrapperClass }}">
+<div class="{{ $fieldWrapperClass }} @error('formProperties.' .  $key){{ $fieldErrorWrapperClass }}@enderror">
     @include('lara-forms-builder::includes.field-label')
     @php
         $fieldValue = $model->$key->value ?? $model->$key;
@@ -138,6 +138,6 @@
             @endif
         </div>
         @include('lara-forms-builder::includes.field-error-message')
-        @include('lara-forms-builder::includes.field-help-text')
     @endif
+    @include('lara-forms-builder::includes.field-help-text')
 </div>