+++ resources/views/form-components.blade.php
@@ -8,7 +8,9 @@
             'readOnly' => $field['readOnly'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
-            'secretValueToggle' => $field['secretValueToggle'] ?? false
+            'secretValueToggle' => $field['secretValueToggle'] ?? false,
+            'fieldModifier' =>  $field['fieldModifier'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('select')
@@ -23,6 +25,7 @@
             'searchable' => $field['searchable'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('radio')
@@ -34,6 +37,7 @@
             'readOnly' => $field['readOnly'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('checkbox')
@@ -44,6 +48,19 @@
             'readOnly' => $field['readOnly'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
+        ])
+        @break
+    @case('yes-no-toggle-switch')
+        @include('lara-forms-builder::components.yes-no-toggle-switch', [
+            'key' => $fieldKey,
+            'label' => $field['label'],
+            'toggleOptions' => $field['options'] ?? ['no' => 'No', 'yes' => 'Yes'],
+            'helpText' => $field['helpText'] ?? '',
+            'readOnly' => $field['readOnly'] ?? false,
+            'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
+            'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('textarea')
@@ -55,6 +72,7 @@
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'rows' => $field['rows'] ?? 5,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('date-picker')
@@ -65,6 +83,7 @@
             'readOnly' => $field['readOnly'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('cards')
@@ -90,6 +109,7 @@
             'readOnly' => $field['readOnly'] ?? false,
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('file')
@@ -102,6 +122,7 @@
             'removeIcon' => $field['removeIcon'] ?? null,
             'preview' => $field['preview'] ?? null,
             'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
         ])
         @break
     @case('search-picker')
@@ -112,6 +133,16 @@
             'placeholder' => $field['placeholder'] ?? '',
             'helpText' => $field['helpText'] ?? '',
             'readOnly' => $field['readOnly'] ?? false,
+            'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
+            'tooltip' => $field['tooltip'] ?? null,
+            'fieldErrorWrapperClass' => $field['field_error_wrapper_class'] ?? $defaultFieldErrorWrapperClasses,
+        ])
+        @break
+    @case('trix-editor')
+        @include('lara-forms-builder::components.trix-editor', [
+            'key' => $fieldKey,
+            'label' => $field['label'],
+            'helpText' => $field['helpText'] ?? '',
             'fieldWrapperClass' => $field['field_wrapper_class'] ?? $defaultFieldWrapperClass,
             'tooltip' => $field['tooltip'] ?? null,
         ])