+++ resources/views/includes/field-label.blade.php
@@ -1,8 +1,7 @@
 <label for="formProperties-{{ $key }}" class="lfb-label">
     {!! $label !!}
-    @if ((!isset($mode) || (isset($mode) and $mode != 'view')) and isset($rules)
-        and array_key_exists('formProperties.' .  $key, $rules) && str_contains($rules['formProperties.' .  $key], 'required'))
-        <sup>*</sup>
+    @if ((!isset($mode) || (isset($mode) and $mode != 'view')))
+        {!! $this->fieldIndicator($key) !!}
     @endif
     @if(isset($tooltip) && $tooltip)
         @include('lara-forms-builder::includes.tooltip', ['message' => $tooltip])