+++ src/LaraFormComponent.php
@@ -4,10 +4,11 @@ namespace WisamAlhennawi\LaraFormsBuilder;
 
 use Livewire\Component;
 use Livewire\WithFileUploads;
+use WisamAlhennawi\LaraFormsBuilder\Traits\FieldIndicator;
 
 abstract class LaraFormComponent extends Component
 {
-    use LaraFormsBuilder, WithFileUploads;
+    use FieldIndicator, LaraFormsBuilder, WithFileUploads;
 
     abstract protected function fields(): array;